remove the sensehat_display_remove function by using devm_add_action
the only thing we do in remove is call `misc_deregister`. If we
use the `devm_add_action` function we can add calling that function
to the list of devres managed actions for this driver that will be
automatically called and cleaned upon unload.

This means that the driver no longer needs a remove function which is nice.
less code => less bugs