tweak logic on ioctl to remove some uneeded work
with the gamma reset ioctl, if the user passes something
invalid, then no change is actually made to the gamma table
and so there is no need to do any sort of update, we can just
unlock the mutex.

even if they do pass something valid, since we know that
the gamme values are coming from the hard coded tables,
there is no need to loop over them again to wrap them mod 32
since we know they are good values.