add children devices to the sensehat device tree
The joystick has its own copy of the gpio information
and both children have their compatible strings and
the smbus registers they use for communication--display
starting at 0x0 and joystick at 0xf2.

Currently this info is not used by the drivers, but
it will be utilized in subsequent patches to simplify
the code and make it more modular and customizable.