complete anihilation of the sensehat core driver
changes in sensehat.h:

The sensehat struct has been entirely removed, the
header now only contains information about the ioctl
for the display driver (maybe it shouldn't go in mfd
anymore?)

changes in sensehat-core.c:

The core driver has been reduced to an absolute stub,
it now only preforms two actions, 1) create the regmap
for the sensehat and 2) call `devm_of_platform_populate`
to create the children devices based on their descriptions
in the device tree. These are the exact actions performed
by the `simple-mfd-i2c` driver upstream and so core will
no longer be included in the patches at all, we will just
add the sensehat compatible to the list of compatible
strings in `simple-mfd-i2c`. The core file is now just here
to facilitate building the sensehat driver out of tree.

changes in sensehat-joystick.c and sensehat-display.c:

To facilitate the switch to matching against compatibles
from the device tree with devm_of_platform_populate, the
module device tables of the two subdrivers needed to be
switched from `platform` to `of` and the compatible strings
needed to be tweaked.